moved common build script to helper
