+++ script/build_swirl.py
@@ -1,12 +1,10 @@
 # Author: Chunyang Wang
 # GitHub Username: chunyang-w
-import csv
-import os
-import shutil
 from argparse import ArgumentParser
 
 import firedrake as fd
 import matplotlib.pyplot as plt
+from build_helper import *
 
 import UM2N
